move SpendNoteInfo and OutputNoteInfo to note.rs
